Drew Node entities on grid

+++ RG_PSI_PZ2/Helpers/CoordinateConversion.cs
@@ -49,9 +49,9 @@ namespace RG_PSI_PZ2.Helpers
             latitude = ((lat + (1 + e2cuadrada * Math.Pow(Math.Cos(lat), 2) - (3.0 / 2.0) * e2cuadrada * Math.Sin(lat) * Math.Cos(lat) * (tao - lat)) * (tao - lat)) * (180.0 / Math.PI)) + diflat;
         }
 
-        public static double Scale(int value, int min, int max, int minScale, int maxScale)
+        public static double Scale(double value, double min, double max, double minScale, double maxScale)
         {
-            return minScale + ((double)(value - min) / (max - min) * (maxScale - minScale));
+            return minScale + ((value - min) / (max - min) * (maxScale - minScale));
         }
     }
 }
